+++ lib/song/utils.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 import '../settings_provider.dart';
-import 'state_provider.dart';
+import 'song_page_state.dart';
 
 List<Widget> getFirstVerseHeader(Book book, String songKey) {
   final List<Widget> firstVerseHeader = [];
